+++ Sources/ProjectSpec/FileType.swift
@@ -72,6 +72,7 @@ extension FileType {
         "bundle": FileType(buildPhase: .resources),
         "xcassets": FileType(buildPhase: .resources),
         "storekit": FileType(buildPhase: .resources),
+        "xcstrings": FileType(buildPhase: .resources),
 
         // sources
         "swift": FileType(buildPhase: .sources),