+++ Sources/XcodeGenKit/XCProjExtensions.swift
@@ -61,6 +61,8 @@ extension Xcode {
             return "wrapper.extensionkit-extension"
         case ("swiftcrossimport", _):
             return "wrapper.swiftcrossimport"
+        case ("xcstrings", _):
+            return "text.json.xcstrings"
         default:
             // fallback to XcodeProj defaults
             return Xcode.filetype(extension: fileExtension)